+++ CHANGES.md
@@ -8,6 +8,11 @@
 
 ### Improvements
 * Made `TokenQueue` and `CharacterReader` autocloseable, to ensure that they will release their buffers back to the buffer pool, for later reuse.
+* Added `Selector#evaluatorOf(String css)`, as a clearer way to obtain an Evaluator from a CSS query. An alias of `QueryParser.parse(String css)`.
+* Custom tags (defined via the `TagSet`) in a foreign namespace (e.g. SVG) can be configured to parse as data tags.
+
+### Bug Fixes
+* The contents of a  `script` in a `svg` foreign context should be parsed as script data, not text. [#2320](https://github.com/jhy/jsoup/issues/2320) 
 
 ## 1.20.1 (2025-04-29)
 
@@ -42,8 +47,8 @@
   Attributes. Also, added `Tag#prefix()`, `Tag#localName()`, `Attribute#prefix()`, `Attribute#localName()`, and
   `Attribute#namespace()` to retrieve these. [#2299](https://github.com/jhy/jsoup/issues/2299).
 * CSS identifiers are now escaped and unescaped correctly to the CSS spec. `Element#cssSelector()` will emit
-  appropriately escaped selectors, and the QueryParser supports those. Added `Selector.escapeCssIdentifier()` and `
-  Selector.unescapeCssIdentifier(). [#2297](https://github.com/jhy/jsoup/pull/2297), [#2305](https://github.com/jhy/jsoup/pull/2305)
+  appropriately escaped selectors, and the QueryParser supports those. Added `Selector.escapeCssIdentifier()` and
+  `Selector.unescapeCssIdentifier()`. [#2297](https://github.com/jhy/jsoup/pull/2297), [#2305](https://github.com/jhy/jsoup/pull/2305)
 
 ### Structure and Performance Improvements